Updated README
This updated README describes the three different aspects of our final project simultaneously.  Because no one single block of these components is in a fully functional form, this README does not describe how to use these blocks of code, but merely describes what they do and where we are taking them next.